4.21 Blogs can only be deleted by the user who added it, returns suitable status code if user doesn't match

+++ part4/controllers/blogs.js
@@ -52,6 +52,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(401).json({
+      error: 'users can only delete their own posts'
+    })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
   // catch not needed because express automatically calls error-handling middleware
